Check Argument mtry

Check Argument mtry

check_mtry(mtry, nv)

Arguments

mtry

Number of covariates to possibly split at in each node. Default is the (rounded down) square root of the number of covariates. Alternatively, one can pass a single-argument function returning an integer, where the argument is the number of covariates.

nv

Number of covariates.

Value

Appropriate value of mtry.