Prediction Method for ocf Objects

Prediction method for class ocf.

# S3 method for ocf
predict(object, data = NULL, type = "response", ...)

Arguments

object

An ocf object.

data

Data set of class data.frame. It must contain at least the same covariates used to train the forests. If data is NULL, then object$full_data is used.

type

Type of prediction. Either "response" or "terminalNodes".

...

Further arguments passed to or from other methods.

Value

Desired predictions.

Details

If type == "response", the routine returns the predicted conditional class probabilities and the predicted class labels. If forests are honest, the predicted probabilities are honest.

If type == "terminalNodes", the IDs of the terminal node in each tree for each observation in data are returned.

References

Di Francesco, R. (2023). Ordered Correlation Forest. arXiv preprint arXiv:2309.08755.

See also

Author

Riccardo Di Francesco

Examples

## Generate synthetic data.
set.seed(1986)

data <- generate_ordered_data(1000)
sample <- data$sample
Y <- sample$Y
X <- sample[, -1]

## Training-test split.
train_idx <- sample(seq_len(length(Y)), floor(length(Y) * 0.5))

Y_tr <- Y[train_idx]
X_tr <- X[train_idx, ]

Y_test <- Y[-train_idx]
X_test <- X[-train_idx, ]

## Fit ocf on training sample.
forests <- ocf(Y_tr, X_tr)

## Predict on test sample.
predictions <- predict(forests, X_test)
head(predictions$probabilities)
#>         P(Y=1)    P(Y=2)     P(Y=3)
#> [1,] 0.3215795 0.4732135 0.20520702
#> [2,] 0.3136190 0.4936867 0.19269430
#> [3,] 0.3624392 0.2213801 0.41618073
#> [4,] 0.3627696 0.4386894 0.19854103
#> [5,] 0.7408662 0.1769811 0.08215271
#> [6,] 0.3670253 0.4038256 0.22914902
predictions$classification
#>   [1] 2 2 3 2 1 2 3 3 3 3 1 1 2 3 3 3 1 1 3 3 2 3 1 1 1 2 3 2 1 3 3 3 2 2 2 3 3
#>  [38] 1 2 3 3 1 1 3 1 1 3 3 2 1 1 1 3 1 3 1 2 1 3 1 2 3 2 3 1 1 3 1 2 2 3 1 1 2
#>  [75] 3 3 1 1 1 2 3 2 3 1 2 3 2 1 2 1 3 3 1 1 2 1 1 3 2 3 2 1 2 1 1 3 1 2 2 1 1
#> [112] 3 1 1 3 3 3 3 2 3 3 2 3 1 3 3 3 3 3 1 3 2 2 2 3 3 1 1 3 2 3 1 1 3 3 2 1 3
#> [149] 3 2 3 1 1 1 2 2 1 2 1 1 3 3 2 3 3 1 3 1 3 3 1 3 2 1 2 3 3 1 2 1 3 1 2 2 3
#> [186] 1 2 1 3 1 2 1 1 3 3 1 1 1 2 3 1 3 2 3 1 1 3 1 3 3 3 1 2 3 3 1 2 1 3 2 1 1
#> [223] 3 1 3 2 1 3 3 1 1 2 1 3 1 3 3 3 3 3 1 3 3 3 3 1 1 1 1 3 2 3 2 1 1 1 2 1 1
#> [260] 2 2 1 2 2 3 2 2 1 3 1 1 1 1 3 2 2 2 3 3 3 1 1 3 1 3 1 3 1 3 1 3 2 2 3 1 1
#> [297] 1 3 2 1 1 1 3 3 1 3 3 2 3 3 3 1 1 2 1 2 3 1 2 1 1 1 3 1 1 3 1 3 1 1 1 1 1
#> [334] 1 2 3 1 1 1 1 3 3 3 3 2 1 3 3 3 1 1 2 1 1 1 1 3 2 1 3 3 3 1 1 1 3 2 2 3 2
#> [371] 1 1 3 2 2 3 1 1 2 1 1 3 3 1 3 1 3 2 2 2 3 3 3 3 2 2 2 1 3 1 1 1 1 3 3 2 2
#> [408] 3 2 3 3 2 1 3 3 3 1 1 1 1 3 1 1 1 3 3 1 1 3 1 2 2 3 3 1 1 3 1 1 3 1 1 1 1
#> [445] 1 3 3 3 1 2 3 3 3 2 3 1 3 3 1 2 3 3 3 1 2 2 2 1 3 2 1 3 1 2 3 3 3 3 1 3 2
#> [482] 1 1 3 2 1 3 1 1 3 2 1 3 1 1 1 3 2 1 3

## Get terminal nodes.
predictions <- predict(forests, X_test, type = "terminalNodes")
predictions$forest.1[1:10, 1:20] # Rows are observations, columns are forests.
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]   90   78  114   46   89   60   51   78   82   102    21    86    34
#>  [2,]   71   57   30   87   48   86   78   53   80   110    99    67    84
#>  [3,]   30   20   95    6   95   45   45   95   55    37    35    61    51
#>  [4,]  104   57   48   86   66   82   60   53   47   112   101    67    87
#>  [5,]   15   25   27    7   45   24   31    7    7     7    16    63    17
#>  [6,]   16   80   64   28   20    3   16   41    7    30    89    21    59
#>  [7,]  100   58   88  114  106   20   82   93   26   111    41    54    40
#>  [8,]   94   66   22    6   77   78   48   36   26    37    54    34    40
#>  [9,]  105   59  110   46  104   70   30   77   26    88    95    32    65
#> [10,]   13   34   69   35   56   31   82   34   71    51    68    92    26
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20]
#>  [1,]    27    82     2    75     6    53   100
#>  [2,]    75    70    68   107    85    95    79
#>  [3,]    38    66    54    71    35    57    38
#>  [4,]    75    70   105    68    65    66    29
#>  [5,]    53    16   105    43    73    15    75
#>  [6,]    51    44    28    30   109     7    42
#>  [7,]    84    24    40    72    93    66    56
#>  [8,]    84    86    61    52    70    41    14
#>  [9,]    92    81    61    20    26    70    40
#> [10,]    66    60    74    53    29    68    68