Marginal Effects for Modified Ordered Random Forests

Nonparametric estimation of marginal effects using an ocf object.

marginal_effects(
  object,
  data = NULL,
  which_covariates = c(),
  eval = "atmean",
  bandwitdh = 0.1,
  inference = FALSE
)

Arguments

object

An ocf object.

data

Data set of class data.frame to estimate marginal effects. It must contain at least the same covariates used to train the forests. If NULL, marginal effects are estimated on object$full_data.

which_covariates

Character vector storing the names of the covariates for which marginal effect estimation is desired. If empty (the default), marginal effects are estimated for all covariates.

eval

Evaluation point for marginal effects. Either "mean", "atmean" or "atmedian".

bandwitdh

How many standard deviations x_up and x_down differ from x.

inference

Whether to extract weights and compute standard errors. The weights extraction considerably slows down the program.

Value

Object of class ocf.marginal.

Details

marginal_effects can estimate mean marginal effects, marginal effects at the mean, or marginal effects at the median, according to the eval argument.

The routine assumes that covariates with more than ten unique values are continuous. Otherwise, covariates are assumed to be discrete.

See also

Author

Riccardo Di Francesco

Examples

## Load data from orf package.
set.seed(1986)

library(orf)
data(odata)
odata <- odata[1:200, ] # Subset to reduce elapsed time.

y <- as.numeric(odata[, 1])
X <- as.matrix(odata[, -1])

## Fit ocf . Use large number of trees.
forests <- ocf(y, X, n.trees = 4000)

## Marginal effects at the mean.
me <- marginal_effects(forests, eval = "atmean")
print(me)
#> ocf marginal effects results 
#> 
#> Data info: 
#> Number of classes:    3 
#> Sample size:          200 
#> 
#> Tuning parameters: 
#> Evaluation:           atmean 
#> Bandwidth:            0.1 
#> Number of trees:      4000 
#> Honest forests:       FALSE 
#> Honesty fraction:     0 
#> 
#> Marginal Effects: 
#>    P'(Y=1) P'(Y=2) P'(Y=3)
#> X1   0.166  -0.227   0.061
#> X2  -0.133   0.049   0.084
#> X3  -0.247   0.038   0.209
#> X4  -0.068  -0.037   0.105
summary(me)
#> ocf marginal effects results 
#> 
#> Data info: 
#> Number of classes:    3 
#> Sample size:          200 
#> 
#> Tuning parameters: 
#> Evaluation:           atmean 
#> Bandwidth:            0.1 
#> Number of trees:      4000 
#> Honest forests:       FALSE 
#> Honesty fraction:     0 
#> 
#> Marginal Effects: 
#>    P'(Y=1) P'(Y=2) P'(Y=3)
#> X1   0.166  -0.227   0.061
#> X2  -0.133   0.049   0.084
#> X3  -0.247   0.038   0.209
#> X4  -0.068  -0.037   0.105

## LATEX.
print(me, latex = TRUE)
#> \begingroup
#>     \setlength{\tabcolsep}{8pt}
#>     \renewcommand{\arraystretch}{1.1}
#>     \begin{table}[H]
#>         \centering
#>         \begin{adjustbox}{width = 0.75\textwidth}
#>         \begin{tabular}{@{\extracolsep{5pt}}l c c c}
#>         \\[-1.8ex]\hline
#>         \hline \\[-1.8ex]
#>         & Class 1 & Class 2 & Class 3 \\
#>         \addlinespace[2pt]
#>         \hline \\[-1.8ex] 
#> 
#>         \texttt{X1} & 0.166 & -0.227 & 0.061 \\ 
#>         \texttt{X2} & -0.133 & 0.049 & 0.084 \\ 
#>         \texttt{X3} & -0.247 & 0.038 & 0.209 \\ 
#>         \texttt{X4} & -0.068 & -0.037 & 0.105 \\ 
#> 
#>         \addlinespace[3pt]
#>         \\[-1.8ex]\hline
#>         \hline \\[-1.8ex]
#>         \end{tabular}
#>         \end{adjustbox}
#>         \caption{Marginal effects.}
#>         \label{table:ocf.marginal.effects}
#>     \end{table}
#> \endgroup

# \donttest{
## Compute standard errors. This requires honest forests.
honest_forests <- ocf(y, X, n.trees = 4000, honesty = TRUE)
honest_me <- marginal_effects(honest_forests, eval = "atmean", inference = TRUE)
honest_me$standard.errors
#>       P'(Y=1)   P'(Y=2)    P'(Y=3)
#> X1 0.13331712 0.2970824 0.10669172
#> X2 0.03562630 0.0886940 0.11742068
#> X3 0.07885076 0.1603431 0.09108637
#> X4 0.02204342 0.0340271 0.03019113
honest_me$p.values # These are not corrected for multiple hypotheses testing!
#>       P'(Y=1)   P'(Y=2)    P'(Y=3)
#> X1 0.09063116 0.8892607 0.08423394
#> X2 0.02433044 0.3057805 0.14517085
#> X3 0.17073321 0.7040726 0.60500832
#> X4 0.98492753 0.7354067 0.71355450

print(honest_me, latex = TRUE)# }
#> \begingroup
#>     \setlength{\tabcolsep}{8pt}
#>     \renewcommand{\arraystretch}{1.1}
#>     \begin{table}[H]
#>         \centering
#>         \begin{adjustbox}{width = 0.75\textwidth}
#>         \begin{tabular}{@{\extracolsep{5pt}}l c c c}
#>         \\[-1.8ex]\hline
#>         \hline \\[-1.8ex]
#>         & Class 1 & Class 2 & Class 3 \\
#>         \addlinespace[2pt]
#>         \hline \\[-1.8ex] 
#> 
#>         \texttt{X1} & -0.226 & 0.041 & 0.184 \\ 
#>                      & (0.133) & (0.297) & (0.107) \\ 
#>         \texttt{X2} & -0.08 & -0.091 & 0.171 \\ 
#>                      & (0.036) & (0.089) & (0.117) \\ 
#>         \texttt{X3} & -0.108 & 0.061 & 0.047 \\ 
#>                      & (0.079) & (0.16) & (0.091) \\ 
#>         \texttt{X4} & 0 & -0.011 & 0.011 \\ 
#>                      & (0.022) & (0.034) & (0.03) \\ 
#> 
#>         \addlinespace[3pt]
#>         \\[-1.8ex]\hline
#>         \hline \\[-1.8ex]
#>         \end{tabular}
#>         \end{adjustbox}
#>         \caption{Marginal effects.}
#>         \label{table:ocf.marginal.effects}
#>     \end{table}
#> \endgroup