Prediction Method for ocf Objects

Prediction method for class ocf.

# S3 method for ocf
predict(object, data = NULL, type = "response", ...)

Arguments

object

An ocf object.

data

Data set of class data.frame. It must contain at least the same covariates used to train the forests. If data is NULL, then object$full_data is used.

type

Type of prediction. Either "response" or "terminalNodes".

...

Further arguments passed to or from other methods.

Value

Desired predictions.

Details

If type == "response", the routine returns the predicted conditional class probabilities and the predicted class labels. If forests are honest, the predicted probabilities are honest.

If type == "terminalNodes", the IDs of the terminal node in each tree for each observation in data are returned.

See also

Author

Riccardo Di Francesco

Examples

## Load data from orf package.
set.seed(1986)

library(orf)
data(odata)
odata <- odata[1:200, ] # Subset to reduce elapsed time.

y <- as.numeric(odata[, 1])
X <- as.matrix(odata[, -1])

## Training-test split.
train_idx <- sample(seq_len(length(y)), floor(length(y) * 0.5))

y_tr <- y[train_idx]
X_tr <- X[train_idx, ]

y_test <- y[-train_idx]
X_test <- X[-train_idx, ]

## Fit ocf on training sample.
forests <- ocf(y_tr, X_tr)

## Predict on test sample.
predictions <- predict(forests, X_test)
head(predictions$probabilities)
#>          P(Y=1)    P(Y=2)     P(Y=3)
#> [1,] 0.56460761 0.2045307 0.23086167
#> [2,] 0.64811163 0.0881887 0.26369967
#> [3,] 0.51968342 0.1830659 0.29725071
#> [4,] 0.11728951 0.1026781 0.78003242
#> [5,] 0.87126267 0.1187196 0.01001771
#> [6,] 0.04049916 0.2181384 0.74136244
predictions$classification
#>   [1] 1 1 1 3 1 3 2 1 1 2 2 3 3 3 1 3 3 2 2 3 3 2 2 3 3 3 2 2 2 1 3 3 3 3 1 2 3
#>  [38] 3 1 3 2 2 1 1 2 3 1 1 3 2 1 1 2 1 1 3 1 3 1 3 1 2 3 3 3 2 3 2 1 1 3 3 3 3
#>  [75] 3 2 1 3 3 1 3 1 3 2 1 1 3 2 1 2 1 3 1 3 3 1 3 1 2 3

## Get terminal nodes.
predictions <- predict(forests, X_test, type = "terminalNodes")
predictions$forest.1[1:10, 1:20] # Rows are observations, columns are forests.
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]   12   10   16   12   15   11    8   13   22    14    13    14     4
#>  [2,]    4    8   10    2   16    2   12    9   16     7    11    14     4
#>  [3,]   22   16   14    8   10   14    6   11   10    12    20    30    22
#>  [4,]    5   12   20    2   18    2   12   10    6    16    10    30    10
#>  [5,]    7   13    5    5    3    5    7   13    7     7     3     3    13
#>  [6,]   10   12   18    2   18    2   12   10   22    16    10    30    10
#>  [7,]   10   12   18   12   20   16   12   11   18    20    20    28    22
#>  [8,]   13   17   18   10   22    5   14   18   16    18    17    22    12
#>  [9,]    4    8   10    2   16    2    6   11   18     6    12     8     4
#> [10,]   19    5   18   10   23   17   14   18    7    18    17    22    12
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20]
#>  [1,]    16    20     7    10    15    17     4
#>  [2,]    15     4     4     5    12    12     7
#>  [3,]    22    16    16    14     6    20    10
#>  [4,]    12    10    16     8    12     9    10
#>  [5,]    23    13     7     5    13    11     7
#>  [6,]    12    10    16     8    12    17    10
#>  [7,]    12    21    18    14    20    14    10
#>  [8,]    23    11    11    11     9    13    11
#>  [9,]    15     4     4    10     6    12     7
#> [10,]    23    17    12    18    18    13    11